kestest: move server to gateway tests
This commit refactors the KES server tests.
Since the KES stateful and stateless server
get separate APIs, this commit refactors the
existing server tests and re-implements them
as gateway tests.

A gateway is truly stateless w.r.t. identities
and policies.
